requirements.txt
# RoboGPT Video Streaming System Dependencies

# Web Framework
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.0

# Video Processing
opencv-python==4.9.0.80
numpy==1.26.0

# Background Tasks
APScheduler==3.10.4

# HTTP Client
requests==2.31.0

# Configuration
PyYAML==6.0.1

# Additional utilities
python-dateutil==2.8.2
